scx_wd40: factor out deadline scheduling logic
